Improve DUX for `create-guest-authors` command
The biggest part of this change is to add a progress bar for `wp co-authors-plus create-guest-authors`.

When doing a mass conversion, it's nice to be able to get a sense of "I should go eat lunch now, this is going to take a while" or "I shouldn't bother switching from Red to 1989, this will be done so quickly I may as well have Red play a third time in a row" or anything in between.  This uses `WP_CLI\Utils\make_progress_bar` to create a progress bar.

Additionally, it adds the count of how many accounts will be attempted at the start in order to further improve the Developer User experience

See http://wp-cli.org/docs/internal-api/wp-cli-utils-make-progress-bar/